Cerebrium/requirements.txt
torch==2.6.0
torchaudio==2.6.0
transformers
accelerate
peft
librosa
silero-vad
chatterbox-tts
aiohttp
aiortc
av
numpy
